+++ src/realm/group_shared.hpp
@@ -369,8 +369,11 @@ public:
     // by compact(). This may not correspond to the space which is free
     // at the point where get_stats() is called, since that will include
     // memory required to hold older versions of data, which still
-    // needs to be available.
-    void get_stats(size_t& free_space, size_t& used_space);
+    // needs to be available. The locked space is the amount of memory
+    // that is free in current version, but being used in still live versions.
+    // Notice that we will always have two live versions - the current and the
+    // previous.
+    void get_stats(size_t& free_space, size_t& used_space, util::Optional<size_t&> locked_space = util::none) const;
     //@}
 
     enum TransactStage {
@@ -576,6 +579,7 @@ private:
 
     // Member variables
     size_t m_free_space = 0;
+    size_t m_locked_space = 0;
     size_t m_used_space = 0;
     Group m_group;
     ReadLockInfo m_read_lock;
@@ -703,9 +707,13 @@ private:
 };
 
 
-inline void SharedGroup::get_stats(size_t& free_space, size_t& used_space) {
+inline void SharedGroup::get_stats(size_t& free_space, size_t& used_space, util::Optional<size_t&> locked_space) const
+{
     free_space = m_free_space;
     used_space = m_used_space;
+    if (locked_space) {
+        *locked_space = m_locked_space;
+    }
 }